Beta 0.5.1
Fog of war is now only revealed in a 9 block radius

forgot to remove a code addition meant only for testing.
(the fog of war not visable for maps)

Fixed start dialog bug on maps.
(not supposed to display "You have laready been here" when you havent
been there)

changed the way the player recieves the key press.
(now recieves the key event, not the key code)